+++ HISTORY.rst
@@ -15,6 +15,9 @@ PlatformIO 2.0
 * Implemented Over The Air (OTA) upgrades for `Espressif <http://docs.platformio.org/en/latest/platforms/espressif.html>`__
   development platform.
   (`issue #365 <https://github.com/platformio/platformio/issues/365>`_)
+* Updated `CMSIS framework <http://docs.platformio.org/en/latest/frameworks/cmsis.html>`__
+  and added CMSIS support for Nucleo F401RE board
+  (`issue #373 <https://github.com/platformio/platformio/issues/373>`_)
 * Added support for Espressif ESP8266 ESP-01-1MB board (ready for OTA)
 * Handle ``upload_flags`` option in `platformio.ini <http://docs.platformio.org/en/latest/projectconf.html>`__
   (`issue #368 <https://github.com/platformio/platformio/issues/368>`_)